# Fix #80: Merge incomplete responses data