Fix https://github.com/devopsfaith/krakend/issues/80 

In commit https://github.com/devopsfaith/krakend/commit/cc67d65263ac836659a0ac190e85fa707de3fba1 there is the test that reproduces the issue, and the fix in https://github.com/devopsfaith/krakend/commit/3882217eb2dbe92ce27d7d227935285df0468217.

I've run the benchmarks, and there is also a minor improvement in speed / allocs:

Before:
```
goos: linux
goarch: amd64
pkg: krakend/proxy
BenchmarkNewMergeDataMiddleware/with_2_parts-8         	  300000	      4593 ns/op	    1360 B/op	      20 allocs/op
BenchmarkNewMergeDataMiddleware/with_3_parts-8         	  300000	      5384 ns/op	    1488 B/op	      22 allocs/op
BenchmarkNewMergeDataMiddleware/with_4_parts-8         	  200000	      5826 ns/op	    1584 B/op	      24 allocs/op
PASS
ok  	krakend/proxy	5.501s
```
After:
```
goos: linux
goarch: amd64
pkg: krakend/proxy
BenchmarkNewMergeDataMiddleware/with_2_parts-8         	  500000	      4339 ns/op	     976 B/op	      17 allocs/op
BenchmarkNewMergeDataMiddleware/with_3_parts-8         	  300000	      5154 ns/op	    1104 B/op	      19 allocs/op
BenchmarkNewMergeDataMiddleware/with_4_parts-8         	  300000	      5789 ns/op	    1200 B/op	      21 allocs/op
PASS
ok  	krakend/proxy	6.747s
```